Renamed files to fit python naming conventions
