+++ Database/Database.py
@@ -1,11 +1,11 @@
 import sqlite3
 from random import randint
 
-from Backend.Employee import Employee
-from Backend.EmployeeAddress import EmployeeAddress
-from Backend.EmployeePermissions import EmployeePermissions
-from Backend.EmployeePTO import EmployeePTO
-from Backend.EmployeeCredentials import EmployeeCredentials
+from Backend.employee import Employee
+from Backend.employee_address import EmployeeAddress
+from Backend.employee_permissions import EmployeePermissions
+from Backend.employee_pto import EmployeePTO
+from Backend.employee_credentials import EmployeeCredentials
 '''
     Database Class
 
@@ -24,7 +24,6 @@ def initialize_employee_database(database : sqlite3.Connection, cursor : sqlite3
 
     with open('Database/EmpDataTriggers.sql') as databaseTriggers:
         database_triggers = databaseTriggers.read().replace('\n', ' ')
-        print(database_triggers)
 
     for command in database_commands:
         cursor.execute(command)
@@ -39,16 +38,17 @@ def initialize_employee_database(database : sqlite3.Connection, cursor : sqlite3
 def generate_employees():
     with open('Database/employees.csv') as employeeFile:
         lineCount = 0
-        for employee in employeeFile:
+        for emp in employeeFile:
             if(lineCount > 0):
-                employee = employee.rstrip().split(',')
-                newEmpAddress = EmployeeAddress(employee[3], employee[4], employee[5], int(employee[6]))
+                emp = emp.rstrip().split(',')
+                newEmpAddress = EmployeeAddress(emp[3], emp[4], emp[5], int(emp[6]))
                 newEmpPermissions = EmployeePermissions(False, False, False, False)
                 newEmpPTO = EmployeePTO(50, 0, 75)
                 newEmpCredentials = EmployeeCredentials(None, None)
                 phoneNum = generate_phone_dashes(randint(2000000000, 9999999999))
                 payType = randint(1, 2)
-                newEmp = Employee(employee[1], employee[2], phoneNum, float(employee[8]), float(employee[10]), float(employee[9]), payType, False, newEmpAddress, newEmpPermissions, newEmpPTO, newEmpCredentials)
+                payMethod = randint(1, 3)
+                newEmp = Employee(emp[1], emp[2], phoneNum, float(emp[8]), float(emp[10]), float(emp[9]), payType, payMethod, False, newEmpAddress, newEmpPermissions, newEmpPTO, newEmpCredentials)
                 newEmp.save()
 
             lineCount += 1
@@ -62,4 +62,3 @@ database = sqlite3.connect('Database/empdata.db')
 cursor = database.cursor()
 
 initialize_employee_database(database, cursor)
-
