--- employee_database.egg-info/PKG-INFO
@@ -1,10 +1,0 @@
-Metadata-Version: 1.0
-Name: employee-database
-Version: 1.0
-Summary: UNKNOWN
-Home-page: UNKNOWN
-Author: UNKNOWN
-Author-email: UNKNOWN
-License: UNKNOWN
-Description: UNKNOWN
-Platform: UNKNOWN
