--- employee_database.egg-info/SOURCES.txt
@@ -1,24 +1,0 @@
-setup.py
-Backend/BackendTest.py
-Backend/Employee.py
-Backend/EmployeeAddress.py
-Backend/EmployeeCredentials.py
-Backend/EmployeePTO.py
-Backend/EmployeePermissions.py
-Backend/__init__.py
-Database/Database.py
-Database/__init__.py
-Gui/__init__.py
-Gui/changePayment.py
-Gui/errorMessage.py
-Gui/lessPTO.py
-Gui/login.py
-Gui/mainScreen.py
-Gui/morePTO.py
-Gui/newEmp.py
-Gui/newPassword.py
-Gui/scrollable.py
-employee_database.egg-info/PKG-INFO
-employee_database.egg-info/SOURCES.txt
-employee_database.egg-info/dependency_links.txt
-employee_database.egg-info/top_level.txt
