--- employee_database.egg-info/top_level.txt
@@ -1,3 +1,0 @@
-Backend
-Database
-Gui
