--- setup.py
@@ -1,3 +1,0 @@
-from setuptools import setup, find_packages
-
-setup(name='employee-database', version='1.0', packages=find_packages())
